pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "flask-migrate>=4.0.7",
    "flask-login>=0.6.3",
    "flask-executor>=1.0.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.2",
    "pytest>=7.0.0",
]
